requirements.txt
asgiref==3.8.1
click==8.1.7
dj-database-url==2.3.0
Django==5.1.2
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
h11==0.14.0
mysqlclient==2.2.5
packaging==24.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
PyMySQL==1.1.1
sqlparse==0.5.1
typing_extensions==4.12.2
uvicorn==0.32.1
whitenoise==6.8.2
